LPS-153827 Move to a more suitable module
